third_party/WebKit/LayoutTests/shadow-dom/imperative-api-notslotchange.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
    const test = async_test('slotchange event must not be fired');
    let eventCount = 0;
    let anothereventCount = 0;
    let  host;
    let slot1;
    let slot2;
    let child1;
    let child2;

    test.step(function () {
        host = document.createElement('div');
        document.body.appendChild(host);

        const shadowRoot = host.attachShadow({mode: 'open', slotting: 'manual'});
        slot1 = document.createElement('slot');

        slot1.addEventListener('slotchange', function (event) {
            eventCount++;
        });

        shadowRoot.appendChild(slot1);

        slot2 = document.createElement('slot');

        slot2.addEventListener('slotchange', function (event) {
            anothereventCount++;
        });

        shadowRoot.appendChild(slot2);

        child1 = document.createElement('div');
        child2 = document.createElement('div');

        host.appendChild(child1);

        slot1.assign([]);
        slot2.assign([child2]);

        assert_equals(eventCount, 0, 'slotchange event must not be fired synchronously');
        assert_equals(anothereventCount, 0, 'slotchange event must not be fired synchronously');
    });

    setTimeout(function () {
        test.step(function () {
            assert_equals(eventCount, 0, 'slotchange event must not be fired  when the slot is assigned none');
            assert_equals(anothereventCount, 0, 'slotchange event must not be fired when the slot is assigned nodes that is not the host\'s child');
            slot1.assign([child1]);
            slot2.assign([]);
        });

        setTimeout(function () {
            test.step(function () {
                assert_equals(eventCount, 1, 'slotchange must be fired exactly once after the assigned nodes changed');
                assert_equals(anothereventCount, 0, 'slotchange must not be fired when deleted nodes that is not the host\'s child');

                slot1.assign([child1]);
                slot2.assign([child1]);
            });

            setTimeout(function () {
                test.step(function () {
                    assert_equals(eventCount, 1, 'slotchange event must not be fired  when the slot is assigned same node');
                    assert_equals(anothereventCount, 0, 'slotchange event must not be fired when it is assigned nodes that is assigned another slot');

                    slot2.assign([]);
                });

                setTimeout(function () {
                    test.step(function () {
                        assert_equals(anothereventCount, 0, 'slotchange event must not be fired when deleted nodes that is assigned another slot');
                    });
                    test.done();
                }, 1);
            }, 1);
        }, 1);
    }, 1);
</script>
</body>
</html>
